+++ threedi_api_client/api.py
@@ -10,6 +10,11 @@ from .auth import refresh_api_key
 from .config import Config, EnvironConfig
 from .versions import API_VERSIONS, host_has_version
 
+import urllib3
+
+
+RETRY_AFTER_STATUS_CODES = frozenset({413, 429, 503, 504})
+
 
 class ThreediApi:
     """Client for the 3Di API.
@@ -104,13 +109,20 @@ class ThreediApi:
         >>> policy = urllib3.util.Retry(total=3, backoff_factor=1.0)
         >>> api = ThreediApi(..., retries=policy)
 
-        For asynchronous usage, you may also supply a ``aiohttp_retry.util.Retry`` object. See
+        For asynchronous usage, you may also supply a ``aiohttp_retry.ExponentialRetry`` object. See
         the `aiohttp_retry docs <https://github.com/inyutin/aiohttp_retry>`_). The ``aiohttp_retry``
         package is shipped with ``threedi_api_client``.
 
         >>> from threedi_api_client.aio import aiohttp_retry
         >>> policy = aiohttp_retry.ExponentialRetry(attempts=3, factor=1.0)
         >>> api = ThreediApi(..., retries=policy)
+
+        Other configuration options are::
+
+        - the exceptions on which to retry (default: None)
+        - the statuses on which to retry (default: 413, 429, 503, 504)
+        - the HTTP methods on which to retry (default: 'DELETE', 'GET', 'HEAD', 'OPTIONS', 'PUT', 'TRACE')
+
     """
 
     def __init__(self, env_file=None, config=None, version="v3", asynchronous=False, retries=3):
@@ -178,6 +190,11 @@ class ThreediApi:
                     retry_options=configuration.retries,
                 )
         else:
+            if isinstance(configuration.retries, int):
+                configuration.retries = urllib3.util.Retry.from_int(configuration.retries)
+            # This adds 504 to the default status codes:
+            if not configuration.retries.status_forcelist:
+                configuration.retries.status_forcelist = RETRY_AFTER_STATUS_CODES
             self._client = ApiClient(configuration)
 
         # Determine what API version to use