+++ HISTORY.rst
@@ -16,7 +16,8 @@ History
 
 - Make V3AlphaApi a subclass of V3BetaApi, and V3BetaApi a superset of V3Api.
 
-- Added a ``rest_config`` to the ``ThreediApi`` which enables setting a retry policy.
+- Added a ``retries`` to the ``ThreediApi`` which enables setting a retry policy.
+  For the async client, ``aiohttp_retry`` is included.
 
 
 3.0.29 (2021-06-02)